I have set up the database (using MONGO DB).
you can generate the small DB using the txt file in the db folder.

I have been set up the express module to split up the server side code to avoid having a huge chunk of code in one file.

I have set up the REST API code for the Ideas. And they are now consuming the mongo db.

The angular js application retrieves the data from the db through the nodejs server.

Next step is to replace the ng router with the UI router and set up the unit tests.